Update layout
Removed excessive parallel routes, added compatibility for expansion
