--- src/app/(home)/(frontpage)/page.tsx
@@ -1,7 +1,0 @@
-export default function Home() {
-  return (
-    <section>
-      
-    </section>
-  );
-}
