--- src/app/(home)/lavalamp/page.tsx
@@ -1,5 +1,0 @@
-export default function Lavalamp() {
-    return (
-        <></>
-    )
-}
