+++ src/app/shrek/page.tsx
@@ -3,7 +3,9 @@ import Shrek from "./Shrek";
 export default function Page() {
 
     return (
-        <Shrek imageUrl={process.env.SHREK_IMAGE} audioUrl={process.env.SHREK_AUDIO} />
+        <body>
+            <Shrek imageUrl={process.env.SHREK_IMAGE} audioUrl={process.env.SHREK_AUDIO} />
+        </body>
     )
 
 }
