+++ src/components/home/tools/dotnet.png
